+++ CHANGES.rst
@@ -39,6 +39,7 @@
 * Changed: Markdown parser has moved to ``coaster.utils.markdown`` and is no
   longer a hack to be embarrassed by.
 * New: Unicode whitespace strippers, ``ulstrip``, ``urstrip`` and ``ustrip``
+* Added lazy role grant evaluation and declarative role grants
 
 
 0.6.0